+++ CodeStarterCamp_Week1/STEP.swift
@@ -1,0 +1,24 @@
+//
+//  main.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by yagom.
+//  Copyright © yagom academy. All rights reserved.
+//
+
+import Foundation
+
+func makeIceCream(iceHeight: Int) {
+    for _ in 1...iceHeight {
+        print("***********")
+                }
+            }
+
+func makeStick(stickHeight: Int){
+    for _ in 1...stickHeight {
+        print("    | |")
+    }
+}
+
+makeIceCream(iceHeight: 8)
+makeStick(stickHeight: 4)
